feat: 디테일 페이지를 홈 페이지 하위로 설정
