+++ src/client/components/Main/MovieList.jsx
@@ -1,37 +1,11 @@
 import React from "react";
-import round from "../../../utils/round";
-import starEmpty from "@images/star_empty.png";
-import { useNavigate } from "react-router-dom";
-import { getThumbnailUrl } from "../../../common/api/tmdb";
+import MovieItem from "./MovieItem";
 
 export default function MovieList({ movies }) {
-  const navigate = useNavigate();
-
   return (
     <ul id="movie-list" className="thumbnail-list">
-      {movies.map(({ id, title, vote_average, poster_path }) => (
-        <li
-          key={id}
-          onClick={() => {
-            console.log(id);
-            navigate(`/detail/${id}`);
-          }}
-        >
-          <div className="item">
-            <img
-              className="thumbnail"
-              src={getThumbnailUrl({ poster_path })}
-              alt={title}
-            />
-            <div className="item-desc">
-              <p className="rate">
-                <img src={starEmpty} className="star" />
-                <span>{round(vote_average, 1)}</span>
-              </p>
-              <strong>{title}</strong>
-            </div>
-          </div>
-        </li>
+      {movies.map((movie) => (
+        <MovieItem key={movie.id} movie={movie} />
       ))}
     </ul>
   );
