+++ src/client/pages/Home.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import Header from "../components/Header";
 import Main from "../components/Main";
 import Footer from "../components/Footer";
-import { useLoaderData } from "react-router-dom";
+import { Outlet, useLoaderData } from "react-router-dom";
 
 export default function HomePage() {
   const movies = useLoaderData();
@@ -12,6 +12,7 @@ export default function HomePage() {
       <Header movie={movies[0]} />
       <Main movies={movies} />
       <Footer />
+      <Outlet />
     </>
   );
 }
