+++ src/common/routes.js
@@ -1,22 +1,25 @@
 import React from "react";
-import DetailPage from "../client/pages/Detail";
+import DetailModal from "../client/pages/DetailModal";
 import HomePage from "../client/pages/Home";
 import { getMovie, getMovies, TMDB_MOVIE_LISTS } from "./api/tmdb";
+
+export const ROUTE = {
+  home: "/",
+  detail: (id) => `/detail/${id}`,
+};
 
 const routes = [
   {
     path: "/",
     loader: async () => await getMovies(TMDB_MOVIE_LISTS.nowPlaying),
     element: <HomePage />,
-  },
-  {
-    path: "/detail/:id",
-    loader: async ({ params }) => {
-      const movies = await getMovies(TMDB_MOVIE_LISTS.nowPlaying);
-      const movie = await getMovie(params.id);
-      return { movies, movie };
-    },
-    element: <DetailPage />,
+    children: [
+      {
+        path: "/detail/:id",
+        loader: async ({ params }) => await getMovie(params.id),
+        element: <DetailModal />,
+      },
+    ],
   },
 ];
